Backend updates

+++ aspnet-core/src/LoanManagement.Application/LoanApplications/Dto/LoanApplicationDto.cs
@@ -1,17 +1,12 @@
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using Abp.Application.Services.Dto;
-using Abp.Authorization.Roles;
-using Abp.AutoMapper;
 using LoanManagement.AssetAndLiablity.Dto;
-using LoanManagement.Authorization.Roles;
 using LoanManagement.BorrowerEmploymentInformations.Dto;
 using LoanManagement.BorrowerInformations;
 using LoanManagement.DetailsOfTransaction.Dto;
 
 namespace LoanManagement.LoanApplications.Dto
 {
-    public class LoanApplicationDto : EntityDto<int>
+    public class LoanApplicationDto : EntityDto<long>
     {
         public MortgageTypeDto MortgageType { get; set; }
         public PropertyInformationDto PropertyInformation { get; set; }
